+++ threedi_modelchecker/checks/factories.py
@@ -9,11 +9,17 @@ from .base import UniqueCheck
 from geoalchemy2.types import Geometry
 
 
+def get_level(table, column, level_map):
+    level = level_map.get(f"*.{column.name}")
+    level = level_map.get(f"{table.name}.{column.name}")
+    return level or "ERROR"
+
+
 def generate_foreign_key_checks(table, custom_level_map=None, **kwargs):
     custom_level_map = custom_level_map or {}
     foreign_key_checks = []
     for fk_column in table.foreign_keys:
-        level = custom_level_map.get(fk_column.parent.name, "ERROR")
+        level = get_level(table, fk_column.parent, custom_level_map)
         foreign_key_checks.append(
             ForeignKeyCheck(
                 reference_column=fk_column.column,
@@ -30,7 +36,7 @@ def generate_unique_checks(table, custom_level_map=None, **kwargs):
     unique_checks = []
     for column in table.columns:
         if column.unique or column.primary_key:
-            level = custom_level_map.get(column.name, "ERROR")
+            level = get_level(table, column, custom_level_map)
             unique_checks.append(UniqueCheck(column, level=level, **kwargs))
     return unique_checks
 
@@ -40,7 +46,7 @@ def generate_not_null_checks(table, custom_level_map=None, **kwargs):
     not_null_checks = []
     for column in table.columns:
         if not column.nullable:
-            level = custom_level_map.get(column.name, "ERROR")
+            level = get_level(table, column, custom_level_map)
             not_null_checks.append(NotNullCheck(column, level=level, **kwargs))
     return not_null_checks
 
@@ -49,7 +55,7 @@ def generate_type_checks(table, custom_level_map=None, **kwargs):
     custom_level_map = custom_level_map or {}
     data_type_checks = []
     for column in table.columns:
-        level = custom_level_map.get(column.name, "ERROR")
+        level = get_level(table, column, custom_level_map)
         data_type_checks.append(TypeCheck(column, level=level, **kwargs))
     return data_type_checks
 
@@ -59,7 +65,7 @@ def generate_geometry_checks(table, custom_level_map=None, **kwargs):
     geometry_checks = []
     for column in table.columns:
         if type(column.type) == Geometry:
-            level = custom_level_map.get(column.name, "ERROR")
+            level = get_level(table, column, custom_level_map)
             geometry_checks.append(GeometryCheck(column, level=level, **kwargs))
     return geometry_checks
 
@@ -69,7 +75,7 @@ def generate_geometry_type_checks(table, custom_level_map=None, **kwargs):
     geometry_type_checks = []
     for column in table.columns:
         if type(column.type) == Geometry:
-            level = custom_level_map.get(column.name, "ERROR")
+            level = get_level(table, column, custom_level_map)
             geometry_type_checks.append(
                 GeometryTypeCheck(column, level=level, **kwargs)
             )
@@ -81,6 +87,6 @@ def generate_enum_checks(table, custom_level_map=None, **kwargs):
     enum_checks = []
     for column in table.columns:
         if issubclass(type(column.type), custom_types.CustomEnum):
-            level = custom_level_map.get(column.name, "ERROR")
+            level = get_level(table, column, custom_level_map)
             enum_checks.append(EnumCheck(column, level=level, **kwargs))
     return enum_checks